// src/types/Habit.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { HabitCategory } from "./HabitCategory";
import type { HabitColor } from "./HabitColor";
import type { HabitStatus } from "./HabitStatus";
import type { HabitStreak } from "./HabitStreak";

export type Habit = {
	id: string;
	name: string;
	description: string;
	created: string;
	color: HabitColor;
	streak: HabitStreak;
	category: HabitCategory;
	status: HabitStatus;
};